Alignment code [Broken]
Still not quite working. I have some stuff commented out that kindof
works, and it worked better earlier on, but I accidentally broke it more
trying to fix the few issues it had. The code for the alignment is
currently in the file "testBed.cpp." I have also transitioned the main
source file to the template required for the competition. The move
function intended for the AI also checks that motor speeds are at a
minimum threshold value if they are supposed to be on, the value is not
forced if the motors are set to 0.